+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 * [bug/5](https://github.com/RishiRamraj/amtk/issues/5): Python 3 support.
 * [bug/4](https://github.com/RishiRamraj/amtk/issues/4): Add examples to the documentation.
+* [bug/14](https://github.com/RishiRamraj/amtk/issues/14): Clean up utils/messages.py so that it can be reused.
 
 ## 0.3.1